pyproject.toml
[tool.poetry]
name = "NowDotAI"
version = "0.2.4"
description = "A lightweight agent interaction framework."
authors = ["NowDotAICrew <github@NowDotAI.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3"
langchain-experimental = "*"
langchain-openai = "*"
python-dotenv = "^1.0.0"
kuzu = "^0.4.0"
cloudpickle = "^3.0.0"
platformdirs = "^4.2.1"
pydantic = "^2.7.1"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
pytest = "^8.0.2"
pytest-cov = "^4.1.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^5.13.2"
sphinx = "^7.3.7"
sphinx-rtd-theme = "^2.0.0"
nbsphinx = "^0.9.4"
pypandoc_binary = "^1.13"
ipykernel = "^6.29.4"
nbsphinx-link = "^1.3.0"
nbformat = "^5.10.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
